Added a lot of documentation. Cleaned up test scripts.
git-svn-id: svn://98.245.80.252/returntome_repository@42 a56a1379-e578-43dd-b219-e15619f59adb

+++ returntome/trunk/Mod/ParseMail.pm
@@ -23,16 +23,19 @@ my $logger = Log::Log4perl->get_logger();
 use Email::Simple;
 
 sub getHeader {
-    my $text = shift;
-    my $header = shift;
-    croak "null argument!" unless ($text && $header);
-    my $email = Email::Simple->new($text);
-    return $email->header($header);
+    my ($text, $header_name)  = @_;
+
+    #Parse the mail:
+    my $email = Email::Simple->new( $text );
+
+    #Get header value:
+    my $header_value = $email->header( $header_name );
+
+    return $header_value;
 }
 
 sub parseMail {
-    my $raw_mail = shift;
-    my $outgoing = shift;
+    my ($raw_mail, $from_address) = @_;
 
     #Create the parser:
     my $parser = new MIME::Parser;
@@ -40,28 +43,35 @@ sub parseMail {
     #Set up the temporary directory the parser will write to:
     my $tmp_dir = '/tmp/mimedump/';
 
-    (-d $tmp_dir) or mkdir $tmp_dir,0755 or die "mkdir: $!"; #TODO: this directory should be permanent
-    rmtree( $tmp_dir, {keep_root => 1} );
-    (-w $tmp_dir) or die "can't write to directory";
+    if ( !(-d $tmp_dir) ) { #if the temp dir doesn't exits, create it
+        mkdir($tmp_dir, 0755) or die "Couldn't create $tmp_dir: $!"; #TODO: this directory should be permanent
+    }
+    else {                  #if it does, empty it
+        rmtree( $tmp_dir, {keep_root => 1} );
+    }
+
+    #Set parser to write to temporary directory:
     $parser->output_dir($tmp_dir);
 
     #parse the mail:
-    my $entity = $parser->parse_data($raw_mail);
+    my $entity = $parser->parse_data( $raw_mail );
 
     #Check the parser for errors:
     my $results  = $parser->results;
     my @msgs = $results->msgs;
-    $logger->error("Parser Error: $_") for (@msgs);    
-    
+    $logger->error("Parser Error: $_") for (@msgs);
+
     #get the headers:
     my $head = $entity->head;
-    my $subject = $head->get('Subject',0);
-    $subject = "\n" unless $subject;
-    my $from = $head->get('From',0);
+    my $subject = $head->get('Subject');
+    if (!$subject) { #check for empty subject line
+        $subject = "\n";
+    }
+    my $from = $head->get('From');
 
     #Initialize mail with basic headers:
     my $parsed_mail = 
-	"From: $outgoing\n" .
+	"From: $from_address\n" .
 	"To: $from" .
 	"Subject: $subject" ;    #TODO: What if subject is encoded? MIME::Head->decode('I_KNOW_WHAT_I_AM_DOING')
 
@@ -70,12 +80,17 @@ sub parseMail {
     my $text_html;
 
     if ($head->count('MIME-Version')) {#Message is MIME formatted
+
 	#Add MIME headers to mail:
 	my @headers = ('MIME-Version','Content-Type','Content-Transfer-Encoding');
-	for my $header (@headers) {
-	    $parsed_mail .= "$header: " . $head->get($header);
+	for my $header_name (@headers) {
+            my $header_value = $head->get($header_name);
+            if ($header_value) {
+                $parsed_mail .= "$header_name: " . $head->get($header_name);
+            }
 	}
 
+        #TODO: rewrite using given-when?
 	switch($entity->effective_type) {
 	    case 'multipart/mixed' { #there are attachments
 	      PARTS:
@@ -92,14 +107,14 @@ sub parseMail {
 			$text_plain = $part;
 			last PARTS;
 		    }
-		}	    
-	    } 
+		}
+	    }
 	    case 'multipart/alternative' { #text/plain and text/html only
 		for my $part ($entity->parts) {
 		    $text_plain = $part if ($part->effective_type eq 'text/plain');
 		    $text_html  = $part if ($part->effective_type eq 'text/html');
 		}
-	    }		
+	    }
 	    case 'text/plain' { #text/plain only
 		$text_plain = $entity;
 	    }
@@ -123,7 +138,7 @@ sub parseMail {
 	    last;
 	}
     }
-    
+
     if  ($instructions) { #instructions were found
 	my ($date, $epoch) = &parseInstructions($instructions);
 	$return_time = $date;
@@ -139,7 +154,7 @@ sub parseMail {
     } else {#instructions were not found, add an error message:
         $error_message = "Sorry, we could not find instructions in this message.";
     }
-    
+
     #Check message size:
     $error_message = "Sorry, your message size must be less than 8 MB." if (length($raw_mail) > 8e6);
 
@@ -155,6 +170,7 @@ sub parseMail {
     #Write the modified message:
     &writeEntity($text_plain,\@plain_lines);
     &writeEntity($text_html,\@html_lines) if $text_html;
+    #TODO: check return value on above
 
     #Add the parsed MIME entity to the mail
     $parsed_mail .= "\n" . join('',@{ $entity->body });
@@ -174,38 +190,59 @@ sub parseMail {
 
 sub readEntity {
     my $entity = shift;
+
     my $bh = $entity->bodyhandle;
 
     my @lines; 
+
+    #Open body handle for reading:
     my $io = $bh->open("r");
     if (not $io) {
 	$logger->error("Could not open MIME entity body");
-	return ();
+	return;
     }
-    while (defined($_ = $io->getline)) { push @lines, $_ }
+
+    #Read from body handle:
+    my $line;
+    while (defined($line = $io->getline)) {
+        push @lines, $line;
+    }
+
+    #Close body handle:
     if (not $io->close) {
 	$logger->error("Could not close MIME entity body");
-	return ();	
-    } 
+	return;
+    }
+
     return @lines;
 }
 
 sub writeEntity {
-    my $entity = shift;
-    my $line_ref = shift;
-    my @lines = @$line_ref;
+    my ($entity, $lines_ref) = @_;
+
+    my @lines = @{ $lines_ref };
 
     my $bh = $entity->bodyhandle;
+
+    #Open body handle for writing:
     my $io = $bh->open("w");
     if (not $io) {
 	$logger->error("Could not open MIME entity body");
 	return 0;
     }
-    $io->print($_) for (@lines);
+
+    #Write lines to body handle:
+    for my $line (@lines) {
+        $io->print($line);
+    }
+
+    #Close body handle:
     if (not $io->close) {
 	$logger->error("Could not close MIME entity body");
-	return 0;	
-    } 
+	return 0;
+    }
+
+    #Return value for success:
     return 1;
 }
 
@@ -214,10 +251,13 @@ use Date::Manip;
 
 sub parseInstructions {
     my $instructions = shift;
-    my $result = ParseDate($instructions);    
+
+    my $result = ParseDate($instructions);
     return unless $result;
+
     my $date = UnixDate($result,"%Y-%m-%d %T");
     my $epoch = UnixDate($result,"%s");
+
     return $date, $epoch;
 }
 
@@ -236,95 +276,241 @@ sub fromEpoch {
     }
 }
 
-
-sub now {
-    return &fromEpoch(time);
-}
-
-
 1;
 
 =head1 NAME
 
-    Mod::ParseMail
-
-=cut
+Mod::ParseMail
 
 =head1 SYNOPSIS
 
-    my %message = %{ &parseMail($raw_message, $from) };
-
-=cut
+C<my %message = %{ &parseMail($raw_message, $from) };>
 
 =head1 DESCRIPTION
 
-    Parses emails.
+Parses emails.
 
-=cut
+=head1 SUBROUTINES
 
-=head1 FUNCTIONS
+=over
 
-=over 
+=item *
 
-=cut
+B<getHeader>
 
-=item getHeader(mail, header name) 
+Extract the specified header from the given email.
 
-    Extract the specified header from the given email.
+I<Arguments:>
 
-=cut
+=over
 
-=item parseMail(raw message, uid)
+=item *
 
-    Extract the return time from the email. 
-    Add error messages or ads as necessary.
-    Arguments: The email as a string, UID, 'From' address
-    Returns: A hashref to a message based on the email.
-    The message looks like:
+mail
+
+=item *
+
+header name
+
+=back
+
+I<Returns:>
+
+=over
+
+=item *
+
+the content of the header
+
+=back
+
+=item *
+
+B<parseMail>
+
+Extract the return time from the email.
+Add error messages or ads as necessary.
+
+I<Arguments:>
+
+=over
+
+=item *
+
+The email as a string
+
+=item *
+
+'From' address
+
+=back
+
+I<Returns:>
+
+=over
+
+=item *
+
+A hashref to a message based on the email.
+The message looks like:
     my %message = (
     mail => $mail,
     return_time => $return_time,
     address => $from,
     );
 
-=cut
+=back
 
-=item readEntity(MIME entity)
-    
-    Get an array of lines representing the body of a MIME::Entity.
-    Arguments: A reference to a MIME::Entity.
-    Returns: A array of lines.
+=item *
 
-=cut
+B<readEntity>
 
-=item writeEntity(entity, lines_ref)
+Get an array of lines representing the body of a MIME::Entity.
 
-    Write an array of lines to the body of a MIME::Entity.
-    Arguments: A reference to a MIME::Entity, a reference to an array of lines.
-    Returns: 1 if write succeeded, 0 if write failed.
+I<Arguments:>
 
-=cut
+=over
 
-=item parseInstructions(instructions)
+=item *
 
-    Extract a return time from the given instructions.
-    Arguments: A string containing the instructions.
-    Returns: Either a time in epoch seconds or undef.
-
-=cut
-
-=item fromEpoch(time)
-
-    Given a time in epoch seconds, return a formatted string representing that time. For safety, the call to DateTime is wrapped in an eval block.
-
-=cut
-
-=item now
-
-    Return a string containing the current time.
-
-=cut
+A reference to a MIME::Entity.
 
 =back
 
-=cut
+I<Returns:>
+
+=over
+
+=item *
+
+A array of lines representing the body of the MIME::Entity.
+
+=back
+
+=item *
+
+B<writeEntity>
+
+Write an array of lines to the body of a MIME::Entity.
+
+I<Arguments:>
+
+=over
+
+=item *
+
+A reference to a MIME::Entity
+
+=item *
+
+A reference to an array of lines.
+
+=back
+
+I<Returns:>
+
+=over
+
+=item *
+
+1 if write succeeded, 0 if write failed.
+
+=back
+
+=item *
+
+B<parseInstructions>
+
+Extract a return time from the given instructions.
+
+I<Arguments:>
+
+=over
+
+=item *
+
+A string containing the instructions.
+
+=back
+
+I<Returns:>
+
+=over
+
+Either a time in epoch seconds or undef.
+
+=back
+
+=item *
+
+B<fromEpoch>
+
+Given a time in epoch seconds, return a formatted string representing that time. For safety, the call to DateTime is wrapped in an eval block.
+
+I<Arguments:>
+
+=over
+
+=item *
+
+Time in epoch seconds.
+
+=back
+
+I<Returns:>
+
+=over
+
+=item *
+
+A formatted string represent the time.
+
+=back
+
+=back
+
+=head1 DEPENDENCIES
+
+=over
+
+=item *
+
+Log::Log4perl
+
+=item *
+
+MIME::Parser
+
+=item *
+
+File::Path
+
+=item *
+
+Mod::Ad
+
+=item *
+
+Switch
+
+=item *
+
+bytes
+
+=item *
+
+File::Path
+
+=item *
+
+Email::Simple
+
+=item *
+
+Date::Manip
+
+=item *
+
+DateTime
+
+=back
